requirements.txt
# ============================================================
# AMD GPU Vehicle Speed Estimation - Requirements
# Python 3.10+ | AMD GPU with DirectML
# ============================================================

# ============================================================
# CORE DEPENDENCIES (Critical Order!)
# ============================================================

# 1. NumPy (MUST be <2.0 for compatibility)
numpy>=1.24.0,<2.0.0

# 2. PyTorch CPU (DirectML installed separately)
torch>=2.0.0
torchvision>=0.15.0

# 3. OpenCV (MUST be <4.10 to avoid NumPy 2.0 issues)
opencv-python>=4.8.0,<4.10.0

# 4. Deep Learning Frameworks
ultralytics>=8.0.0          # YOLOv8/v11
supervision>=0.18.0         # Tracking & Visualization

# ============================================================
# CONFIGURATION & I/O
# ============================================================
pyyaml>=6.0
python-dotenv>=1.0.0

# ============================================================
# PROGRESS & LOGGING
# ============================================================
tqdm>=4.65.0
colorama>=0.4.6             # Colored console output
loguru>=0.7.0               # Advanced logging

# ============================================================
# DATA PROCESSING
# ============================================================
pandas>=2.0.0
scikit-learn>=1.3.0

# ============================================================
# IMAGE & VIDEO PROCESSING
# ============================================================
pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# ============================================================
# VISUALIZATION
# ============================================================
matplotlib>=3.7.0
seaborn>=0.12.0

# ============================================================
# UTILITIES
# ============================================================
scipy>=1.11.0

# ============================================================
# AMD GPU SUPPORT (torch-directml)
# ============================================================

# ⚠️  IMPORTANT: torch-directml has strict version requirements!
# 
# Option 1: Manual installation (RECOMMENDED)
# ───────────────────────────────────────────────────────────
# After installing all packages above, run:
#   pip install torch-directml --no-cache-dir
#
# Option 2: Include in requirements (ADVANCED)
# ───────────────────────────────────────────────────────────  
# Uncomment the line below ONLY if you understand dependency conflicts:
# torch-directml>=0.2.0
#
# WHY separate installation?
# ─────────────────────────
# 1. torch-directml requires specific PyTorch version
# 2. pip resolver may install wrong torch version if all together
# 3. NumPy version conflicts (directml needs <2.0)
# 4. Better control over installation order
# 5. Easier troubleshooting if installation fails
#
# For automatic installation, use setup.bat instead!

# ============================================================
# INSTALLATION INSTRUCTIONS
# ============================================================
# 
# METHOD 1: Automatic (Recommended)
# ──────────────────────────────────
#   setup.bat
#
# METHOD 2: Manual Step-by-Step
# ──────────────────────────────────
# 1. Create virtual environment:
#    python -m venv venv
#    venv\Scripts\activate
#
# 2. Install NumPy first (CRITICAL):
#    pip install "numpy>=1.24.0,<2.0.0"
#
# 3. Install PyTorch CPU:
#    pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
#
# 4. Install OpenCV (BEFORE ultralytics):
#    pip install "opencv-python>=4.8.0,<4.10.0"
#
# 5. Install remaining packages:
#    pip install -r requirements.txt
#
# 6. Install DirectML (AMD GPU):
#    pip install torch-directml --no-cache-dir
#
# OR simply run: setup.bat
